adding utility file for common functionality in system/templates/base/src/js
