+++ system/templates/base/src/js/components/MultiFileUpload.ts
@@ -1,16 +1,6 @@
 // src/js/components/MultiFileUpload.ts
 
-function formatBytes(bytes, decimals = 2) {
-  if (bytes === 0) return '0 Bytes';
-
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-}
+import { formatBytes } from "../utility";
 
 export class MultiFileUpload {
     private static containerTarget = 'multi-upload-file-container';
